Under WSL run win32 apm.exe
When running Linux under WSL on a Windows host, select apm.exe
for local execution. WSL is able to execute win32 binaries, and
so able to use those features (e.g. IPOPT) that are excluded from
the Linux version of the apm binary. Update README.md accordingly.

I do not have access to an arm machine to verify that WSL for arm
is also able to run win32 binaries or that its emulated performance
is acceptable; therefore, the choice to order WSL check after the
arm machine check is conservatively deliberate.

The chosen method of WSL detection follows the instructions from
Microsoft in a manner similar to how it's done by systemd:

 - https://github.com/microsoft/WSL/issues/423#issuecomment-221627364
 - https://github.com/systemd/systemd/blob/a7e508f/src/basic/virt.c#L473